parsing fixed as per database-changes

+++ API/Model/Answer/AnimeAnswer.cs
@@ -24,11 +24,11 @@ namespace AniSharp.API.Model.Answer
                 kanjiName       = x.asString(4),
                 englishName     = x.asString(5),
                 otherName       = x.asString(6),
-                episodes        = x.asInt16(7),
-                highestNoEp     = x.asInt16(8),
-                specialEpCount  = x.asInt16(9),
-                rating          = x.asInt16(10),
-                tempRating      = x.asInt16(11)
+                episodes        = x.asInt(7),
+                highestNoEp     = x.asInt(8),
+                specialEpCount  = x.asInt(9),
+                rating          = x.asInt(10),
+                tempRating      = x.asInt(11)
             };
         }
     }
